Step R17: move LazyLoadProxy to Tier B, and fix what it found (#56)
The move deletes 700 lines and adds 68. The Tier A class ignored `Milk` and
`Culture` on twenty-nine entity types, because InMemory has no complex types,
and then carried two 680-line JSON strings restating the model that was left.
The relational base maps both complex properties itself, so all of it goes.

EF's own SQLite class differs from the core base by one token across both
strings: `"Charge": 1.00` becomes `1.0`, because SQLite has no decimal type and
the scale that comes back is the scale the store wrote. Written as the
substitution rather than as 1360 lines of duplicated JSON.

Four tests then failed with `InvalidCastException: Double to Int32`, one defect.
`ClientResultMaterializer.Materialize` fills the value buffer by bare property
name, and a complex leaf's `Name` is bare: `Host.Culture.Rating` is `"Rating"`.
`Host.Rating` is a `double` and the four `Rating` properties under `Culture` and
`Milk` are `int`, so all four slots took the same double. The method's own remark
already warned that a complex value must not ride in that dictionary because the
names collide; the leak in the other direction was not considered. Only a
property whose `DeclaringType is IEntityType` now reads from the dictionary.

26896 / 26545 / 116 / 235 (`groupc-lazyloadproxy`), unchanged from the baseline.
Trim ratchet 89 <= 89. `CI=true` Release build clean.
